requirements.txt
pandas==2.2.3
Pillow==10.4.0
Requests==2.32.3
scikit_learn==1.5.2
streamlit==1.38.0
streamlit_lottie==0.0.5
